TSS-1482: Barrier Downloads (#714)
* init

* refactor

* Update

* Presigned URL view

* Updates + tests

* format

* format

* fixt dataset test

* remaining tests

* format

* Patch + additional tests

* format

* Rename barrier-reports -> barrier-downloads

* remaining report name changes

* format

* updates

* format

* returned a paginated list to match frontend signature for  a list result

* fix test

* format

* f

* delete download

* barrier fiters

* format

* format;

* f

* id return name

* adding count to database

* tests

* formatg

* Reinstate original qs

* f

* format

* fixing filters added to search when download is saved

* fixing broken test

---------
