+++ api/barrier_downloads/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class BarrierDownloadsConfig(AppConfig):
+    name = "api.barrier_downloads"
